Add issue templates (#41)
* doc: update the documentation;

* doc: update the documentation;

* refactor: preprocessing functions of specific dataset now move to module load_preprocessing;

* fix: solve the problem of circular import;

moved the functions of parsing delta to util.py.

* refactor: don't save data into h5 files if the datasets already exit;

* feat: add issue templates of bug report, feature request, and model addition;
