+++ .github/ISSUE_TEMPLATE/bug-report.yml
@@ -1,0 +1,39 @@
+name: "🐞 Bug Report"
+description: Submit a bug report to help us improve PyPOTS
+body:
+  - type: textarea
+    id: system-info
+    attributes:
+      label: System Info
+      description: Please share information about your operating environment with us.
+      placeholder: PyPOTS version, Python version, platform and your OS version, and versions of other libraries existing in the error message...
+    validations:
+      required: true
+
+  - type: checkboxes
+    id: information-scripts-examples
+    attributes:
+      label: Information
+      description: 'The problem arises when running:'
+      options:
+        - label: "The official example scripts"
+        - label: "My own modified scripts"
+
+  - type: textarea
+    id: reproduction
+    validations:
+      required: true
+    attributes:
+      label: Reproduction
+      description: |
+        Please provide a code sample that reproduces the problem you ran into. It can be a Colab link or just a code snippet.
+        If you have code snippets, error messages, stack traces please provide them here as well.
+        Important! Use code tags to correctly format your code. See https://help.github.com/en/github/writing-on-github/creating-and-highlighting-code-blocks#syntax-highlighting
+        Do not use screenshots, as they are hard to read and (more importantly) don't allow others to copy-and-paste your code.
+
+      placeholder: |
+        Steps to reproduce the behavior:
+          
+          1.
+          2.
+          3.
