+++ .github/ISSUE_TEMPLATE/feature-request.yml
@@ -1,0 +1,31 @@
+name: "🙌 Feature request"
+description: Submit a request to add a new feature
+labels: [ "New feature" ]
+
+body:
+  - type: textarea
+    id: feature-description-requested
+    validations:
+      required: true
+    attributes:
+      label: Feature description
+      description: |
+        A clear and concise description of the feature proposal. Please provide a link to the paper and code if they exist.
+
+  - type: textarea
+    id: motivation
+    validations:
+      required: true
+    attributes:
+      label: Motivation
+      description: |
+        Please outline the motivation of your proposal. Is your feature request related to a problem? If this is related to another GitHub issue, please link here.
+
+  - type: textarea
+    id: contribution
+    validations:
+      required: true
+    attributes:
+      label: Your contribution
+      description: |
+        Is there any way that you could help, for example, by submitting a PR?
